+++ rpsection.pas
@@ -1653,8 +1653,8 @@ begin
  begin
   cachedpos:=-1;
   FirstPage:=0;
-  FDecompStream.SetSize(0);
-  FOldStream.SetSize(0);
+  FDecompStream.SetSize(Int64(0));
+  FOldStream.SetSize(Int64(0));
  end;
 end;
